+++ development/Sandbox/Program.cs
@@ -9,7 +9,6 @@ Console.OutputEncoding = Encoding.UTF8;
 await Sandbox.Providers.AnthropicDemo.RunAsync();
 //await Sandbox.Providers.AmazonBedrock.RunAsync();
 //await Sandbox.Providers.OpenAI.RunAsync();
-//await Sandbox.Providers.GitHub.RunAsync();
 //await AzureOpenAI.RunAsync();
 //await Sandbox.Providers.Mistral.RunAsync();
 //await Sandbox.Providers.Google.RunAsync();